chore: fix icons
